Add solution for reverse compliment problem
This one is 6x faster than my previous Ruby implementation. I was curious if I
could translate the string using `tr`, but the `system` function doesn't return
the input to awk. Interesting limitation; you could probably redirect the output
to a file or pipe and then read it back in.