Add solution to the GC content problem
This one is 11.75x faster than my previous Ruby solution, continuing the trend.
I'm not entirely happy with this solution since it's memory bound, I'd like to
refactor it to compare the max value on each sequence, but the sequences are
broken across multiple lines, so I'd need to detect sequences as a block, which
is a bit tricky.