Add disclaimer to README
I noticed that the Rosalind Project actually discourages posting solutions online. I'm not sure if I want to make this repo private, since I plan to use some of my research here as material for a series of blog posts about awk. I may look into ways of keeping the solutions private if it seems like there is a problem.